backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.5
pydantic==2.10.6
pydantic-settings==2.7.1
orjson==3.10.15
regex==2024.11.6
torch==2.5.1
transformers==4.48.1
spacy==3.8.5
pdfplumber==0.11.5
python-docx==1.1.2
